- STM32F4, STM32F7 시리즈 구조 변경 작업 중 - EFM32, EFR32 시리즈 등록 중

+++ inc/util/runtime.h
@@ -26,7 +26,7 @@
 namespace runtime
 {
 uint32_t getSec(void);
-uint32_t getMsec(void);
+uint64_t getMsec(void);
 
 #if !(defined(__CORE_CM0PLUS_H_GENERIC) || defined(__CORE_CM0_H_GENERIC))
 uint64_t getUsec(void);
